Правки наставника
