+++ controllers/RegisterController.php
@@ -6,7 +6,7 @@ namespace app\controllers;
 use yii\web\Controller;
 use app\models\Users;
 use app\models\Register;
-use app\models\Cities;
+use app\models\City;
 use Yii;
 
 class RegisterController extends Controller
@@ -15,7 +15,10 @@ class RegisterController extends Controller
     {
 
         $register = new Register();
-        $cities = Cities::find()->all();
+
+        //  $cities = \app\models\Cities::find()->select('city')->indexBy('city')->column();
+
+        $cities = City::find()->asArray()->all();
 
         if ($register->load(\Yii::$app->request->post()) & $register->validate()) { // если данные проходят валидацию то делаем дальше
 
@@ -29,11 +32,9 @@ class RegisterController extends Controller
 
             if ($users->save()) {
                 return Yii::$app->response->redirect(['tasks']); // если успех то на страницу с тасками
-            } else {
-                var_dump($users->getErrors());
             }
         } else {
-            return $this->render('index', ['register' => $register]); // а если нет то на ту же страницу попадаем
+            return $this->render('index', ['register' => $register, 'cities' => $cities]); // а если нет то на ту же страницу попадаем
         }
