+++ controllers/TasksController.php
@@ -2,11 +2,11 @@
 
 namespace app\controllers;
 
-use app\models\Categories;
+use app\models\Category;
 use app\models\Replies;
 use app\models\SearchTasks;
 use yii\web\Controller;
-use app\models\Tasks;
+use app\models\Task;
 use yii\web\NotFoundHttpException;
 
 
@@ -18,7 +18,7 @@ class TasksController extends Controller
 
         $modelSearch = new SearchTasks();
         $dataProvider = $modelSearch->search($this->request->post());
-        $categories = Categories::find()->where([])->all();
+        $categories = Category::find()->all();
 
         return $this->render('index', [
             'modelSearch' => $modelSearch,
@@ -31,7 +31,7 @@ class TasksController extends Controller
     public function actionView($id)
     {
 
-        $task = Tasks::find()->where(['id' => $id])->one();
+        $task = Task::find()->where(['id' => $id])->one();
         if ($task === null) {
             throw new NotFoundHttpException("Нету такого таска!");
         }
