+++ controllers/UserController.php
@@ -3,11 +3,11 @@
 namespace app\controllers;
 
 use app\models\UserRating;
-use app\models\UserReplies;
-use app\models\Users;
+use app\models\UserReply;
+use app\models\User;
 use yii\web\Controller;
-use app\models\Tasks;
-use app\models\Replies;
+use app\models\Task;
+use app\models\Reply;
 use Yii;
 use yii\web\NotFoundHttpException;
 
@@ -15,7 +15,7 @@ class UserController extends Controller
 {
     public function actionView($id)
     {
-        $singleUser = Users::find()->where(['id' => $id])->one();
+        $singleUser = User::find()->where(['id' => $id])->one();
         if ($singleUser === null) {
             throw new NotFoundHttpException("Нету такого юзера!");
         }
@@ -24,7 +24,7 @@ class UserController extends Controller
             throw new NotFoundHttpException("Юзер не является исполнителем!");
         }
 
-        $numberReplies = UserReplies::find()->where(['executor_id' => $id])->count();
+        $numberReplies = UserReply::find()->where(['executor_id' => $id])->count();
 
         return $this->render('index',
             [
