+++ models/User.php
@@ -2,108 +2,192 @@
 
 namespace app\models;
 
-use yii\db\ActiveRecord;
-use \yii\web\IdentityInterface;
+use Yii;
 
-class User extends ActiveRecord implements \yii\web\IdentityInterface
+/**
+ * This is the model class for table "users".
+ *
+ * @property int $id
+ * @property string $email
+ * @property string $name
+ * @property string $password_hash
+ * @property int|null $dt_add
+ * @property string|null $user_img
+ * @property string|null $quote
+ * @property string|null $country
+ * @property string $city
+ * @property string|null $age
+ * @property string|null $phone
+ * @property string|null $telegram
+ * @property int|null $status
+ * @property string|null $user_status
+ * @property int|null $answer_orders
+ */
+class User extends \yii\db\ActiveRecord
 {
-    public $id;
-    public $username;
-    public $password;
-    public $authKey;
-    public $accessToken;
-
-    private static $users = [
-        '100' => [
-            'id' => '100',
-            'username' => 'admin',
-            'password' => 'admin',
-            'authKey' => 'test100key',
-            'accessToken' => '100-token',
-        ],
-        '101' => [
-            'id' => '101',
-            'username' => 'demo',
-            'password' => 'demo',
-            'authKey' => 'test101key',
-            'accessToken' => '101-token',
-        ],
-    ];
-
-
     /**
      * {@inheritdoc}
      */
-    public static function findIdentity($id)
+    public static function tableName()
     {
-        return isset(self::$users[$id]) ? new static(self::$users[$id]) : null;
+        return 'user';
     }
 
     /**
      * {@inheritdoc}
      */
-    public static function findIdentityByAccessToken($token, $type = null)
+    public function rules()
     {
-        foreach (self::$users as $user) {
-            if ($user['accessToken'] === $token) {
-                return new static($user);
+        return [
+            [['email', 'name', 'password_hash', 'city'], 'required'],
+            [['dt_add', 'status', 'answer_orders'], 'integer'],
+            [['email', 'name', 'password_hash', 'user_img', 'country', 'city', 'age', 'phone', 'telegram', 'user_status'], 'string', 'max' => 255],
+            [['quote'], 'string', 'max' => 1000],
+        ];
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function attributeLabels()
+    {
+        return [
+            'id' => Yii::t('app', 'ID'),
+            'email' => Yii::t('app', 'Email'),
+            'name' => Yii::t('app', 'Name'),
+            'password_hash' => Yii::t('app', 'Password Hash'),
+            'dt_add' => Yii::t('app', 'Dt Add'),
+            'user_img' => Yii::t('app', 'User Img'),
+            'quote' => Yii::t('app', 'Quote'),
+            'country' => Yii::t('app', 'Country'),
+            'city' => Yii::t('app', 'City'),
+            'age' => Yii::t('app', 'Age'),
+            'phone' => Yii::t('app', 'Phone'),
+            'telegram' => Yii::t('app', 'Telegram'),
+            'status' => Yii::t('app', 'Status'),
+            'user_status' => Yii::t('app', 'User Status'),
+            'answer_orders' => Yii::t('app', 'Answer Orders'),
+        ];
+    }
+
+    /**
+     * {@inheritdoc}
+     * @return Users the active query used by this AR class.
+     */
+
+    public function getAvgRating()
+    {
+        static $rating = null;
+
+        if (is_null($rating) && count($this->replies) > 0) {
+            $ratings = [];
+            foreach ($this->replies as $reply) {
+                $ratings[] = $reply->rate;
             }
-        }
 
-        return null;
+            $rating = array_sum($ratings) / count($ratings);
+        }
+        return $rating;
     }
 
-    /**
-     * Finds user by username
-     *
-     * @param string $username
-     * @return static|null
-     */
-    public static function findByUsername($username)
+    public function getUserAvgRating()
     {
-        foreach (self::$users as $user) {
-            if (strcasecmp($user['username'], $username) === 0) {
-                return new static($user);
+        static $rating = null;
+
+        if (is_null($rating)) {
+            $ratings = [];
+            foreach ($this->executorReplies as $reply) {
+                $ratings[] = $reply->rate;
             }
+
+            $rating = array_sum($ratings) / count($ratings);
+
+
         }
-
-        return null;
+        return $rating;
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getId()
+    public function getReplies()
     {
-        return $this->id;
+        return $this->hasMany(UserReply::class, ['user_id' => 'id']);
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getAuthKey()
+    public function getDoneTasks()
     {
-        return $this->authKey;
+        return $this->hasMany(Task::class, ['user_id' => 'id'])->andWhere(['status' => 1]);
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function validateAuthKey($authKey)
+    public function getFailedTasks()
     {
-        return $this->authKey === $authKey;
+        return $this->hasMany(Task::class, ['user_id' => 'id'])->andWhere(['status' => 0]);
     }
 
-    /**
-     * Validates password
-     *
-     * @param string $password password to validate
-     * @return bool if password provided is valid for current user
-     */
-    public function validatePassword($password)
+    public function getExecutorReplies()
     {
-        $this->password =$password;
-        return \Yii::$app->security->validatePassword($password, $this->password);
+        return $this->hasMany(UserReply::class, ['executor_id' => 'id']);
+    }
+
+    public function getUserRating()
+    {
+        // отношение к новой таблице отзывов пользователя
+        return $this->hasMany(UserRating::class, ['user_id' => 'id']);
+    }
+
+    public function getTags()
+    {
+        return $this->hasMany(TagsAttribute::class, ['id' => 'id'])->viaTable('tags_attribution', ['user_id' => 'id']);
+    }
+
+    public function getCities()
+    {
+        return $this->hasOne(City::class, ['id'=>'city_id']);
+
+    }
+
+
+    public function getAllRepliesForUsers()
+    {
+        return $this->hasMany(TasksReplies::class, ['user_id' => 'id'])->viaTable('replies_links', ['replies_id' => 'id'])->where(['entity' => 'user']);
+    }
+
+    public function getUser()
+    {
+        return $this->hasOne(UserReply::class, ['executor_id' => 'id']);
+    }
+
+    public function getRating()
+    {
+        $id = \Yii::$app->request->get('id');
+
+        /*  $rating = Yii::$app->db->createCommand("
+          SELECT * FROM (SELECT *, (@position:=@position+1) as rate
+          FROM (SELECT executor_id,
+          SUM(rate) / COUNT(rate) as pts
+          FROM user_replies, (SELECT @position:=0) as a
+          GROUP BY executor_id
+          ORDER BY pts DESC) AS subselect)
+          as general WHERE  executor_id = $id")->queryOne();
+  */
+        $rating = '';
+        $cache = Yii::$app->cache;
+// Формируем ключ
+        $key = 'rating';
+// Пробуем извлечь категории из кэша.
+        $rating = $cache->get($rating);
+        if ($rating === false) {
+            //Если $categories нет в кэше, вычисляем заново
+            $rating = Yii::$app->db->createCommand("
+        SELECT * FROM (SELECT *, (@position:=@position+1) as rate 
+        FROM (SELECT executor_id,
+        SUM(rate) / COUNT(rate) as pts
+        FROM user_reply, (SELECT @position:=0) as a
+        GROUP BY executor_id
+        ORDER BY pts DESC) AS subselect) 
+        as general WHERE  executor_id = $id")->queryOne();
+            // Сохраняем значение $categories в кэше по ключу. Данные можно получить в следующий раз.
+            $cache->set($key, $rating);
+        }
+        return $rating;
 
     }
 }
